Rename uint/int to usize/isize, and make usize the default integer type.
Prior to this change, Hemlock's int was explicitly equivalent to OCaml's int.
Now Hemlock's usize is explicitly equivalent to OCaml's int, and isize is
opaque.  This means that usize is now the Hemlock integer type which can be
created via literal constants and matched with literal constants; isize now
requires Isize.kv for constant creation, and extra work to deal with in pattern
matches.

Care is required in the narrow unsigned integer modules (U21 and U8), because
within those modules the compiler knows that e.g. u21 and usize are equivalent
types.  Thus it is possible to create out-of-range constants if kv calls are
omitted.

Remove the Intable interface.  It was used only by Bool and Float.  Bool now
has {to,of}_usize, and Float now has {to,of}_isize.

Fix integer printing and to_string for narrow integers by consolidating pp,
pp_x, and to_string into Intnb.Make_common.