Implement MurmurHash3-based hashing.
Implement hashing based on the 128-bit variant of MurmurHash3.  This requires
the addition of u128 for API purposes.  u128 in turn builds on top of u64, i64,
and u32.

Add bit_ssr to unsigned integer types, because behavior is well defined, and the
function has legitimate uses.

Expand testing of integer types and fix various latent bugs.
- Floating point conversion was badly broken in many cases.  Improve it enough
  to allow correct implementation for u128, but don't bother covering up OCaml's
  unfortunate behavior on overflow for intnb-based integer types.
- Fix edge condition handling for {floor,ceil}_{pow2,lg}.

Resolve #8.
Resolve #39.