Implement/enhance hash_fold for various containers.
Implement hash_fold for List, Option, Set, and Ordset.

Interpose hashes of element indices and elements, followed by container length,
to minimize the potential for spurious collisions.  For example, hashing the
following sequences results in different hashes under the interposition regime:

  let rec hash_seq seq state = begin
    match seq with
    | [] -> state
    | l :: seq' ->
      state |> List.hash_fold Usize.hash_fold l |> hash_seq seq'
  end in
  let seqs = [
    [];
    [[]];
    [[]; []];

    [[0]];
    [[0]; []];
    [[]; [0]];
    [[]; []; [0]];
  ] in
  List.iter seqs ~f:(fun seq ->
    Format.printf "@[<h>%a@]\n" Hash.State.pp (hash_seq seq Hash.State.seed)
  )

Enhance Array to use the interposition mechanism as described above.

Enhance String and Bytes to finish with the hash of the byte length.  This is
sufficient to prevent no-op hashing of empty collections; interposition serves
no purpose for monomorphic collections.

Resolve #51.