Optimize Ordmap join/split, and avoid split wherever possible.
Refactor join to build node structures directly, rather than creating temporary
nodes and rotating them.

Optimize split_node and split2_node to avoid building nodes that are identical
to those already in the tree, e.g. when the split key precedes/follows the
entire tree.

Fix a bug in split2_node: do not omit the nodes in the path to the split node.
This affected the disjoint function.

Optimize {insert,update,remove}[_hlt], upsert, and amend to avoid split.