Enable multi-platform Docker build.
This enables us to manually set the target platform for our image using the HEMLOCK_PLATFORM
environment variable. There are less-verbose ways of plumbing the target platform through, such as
using the `platform: ` key in docker-compose.yml. However, by doing it in here, the default value
for HEMLOCK_PLATFORM is set equal to the docker buildx BUILDPLATFORM environment variable, which
matches the builder's host platform. That means that in the cases where we connect a remote builder
(AWS Graviton, for instance), the default target platform will match that of the builder and not of
the local host where the CLI commands are invoked.