Move CI testing to client.
For pull requests, we now run testing on the client side. GitHub actions does
not currently provide a new enough kernel to run Hemlock's File I/O with
io_uring, so we require that tests run elsewhere.  After pushing a branch,
status for the latest commit in the branch should be set to either `success` or
`failure` with a context of `Hemlock Push` and a description ending in `tree
<sha>` (the output of `git cat-file -p HEAD | head -n 1`). Our GitHub "Test"
action no longer runs tests. Instead, it verifies that the client ran the tests,
that the tests succeeded, and that the branch was correctly based on
`origin/master`.

We acheive all this by using our custom `gh` CLI extension at
https://github.com/BranchTaken/gh-push/. Use the provided `gh push` command to
run tests, push branches, and set commit status.