Enhance Source.Slice.line_context to support source directives
Line context for a source slice may span multiple sources, and it should be up
to the caller to deal with this scenario. Therefore line context now comprises a
non-empty list of source slices.

Fix various subtle flaws in bias chaining that line context testing exposed,
both in Scan and Source.