Merge `Range[HF]` into `Range`
Rather than having separate types for half-open and full-open ranges, merge the
two such that each range value tracks whether it is half-open or full-open. This
slightly complicates logic related to length and limit (past/last), but most
uses of ranges are based on cursors and slices, so the extra complications do
not routinely affect application code.