Implement `-[re]m[erge]` command line option
Equivalent state remerging can obscure conflict attribution due to reporting
unions of attribs that may not be identical. Disabling remerging is only useful
as a tool during grammar development (no point in using it for production
parsers), but it can be important for understanding how inadequacies impact the
generated parser.